Refactor: shared authenticated `octokit` instance
